+++ src/authorisation_token_scraper/api/apple/podcasts.py
@@ -1,4 +1,4 @@
-from authorisation_token.scraper import Scraper
+from authorisation_token_scraper.scraper import Scraper
 
 
 class Podcasts(Scraper):
